Implements user registration functionality
Adds a registration form with validation, including handling for
"Divers" gender selection. It also includes logic to submit the
registration data to the backend and redirect to the login page
upon successful registration.

Refactors the login process to use the email as username.
Also ensures password values are not trimmed during login.

Introduces AuthStorageService and UserStorageService for managing
authentication tokens and user data in local storage.
These services provide methods for setting, clearing, and retrieving
tokens and user information.